+++ static/src/javascripts/bootstraps/enhanced/profile.js
@@ -18,7 +18,6 @@ import { initUserEditLink } from 'common/modules/discussion/user-edit-link';
 import { init as initTabs } from 'common/modules/ui/tabs';
 import { enhanceAdPrefs } from 'common/modules/identity/ad-prefs';
 import { enhanceUpsell } from 'common/modules/identity/upsell/upsell';
-import { enhanceFollow } from 'common/modules/identity/follow';
 
 const initFormstack = (): void => {
     const attr = 'data-formstack-id';
@@ -67,7 +66,6 @@ const initProfile = (): void => {
         ['enhance-consent-journey', enhanceConsentJourney],
         ['init-header', initHeader],
         ['init-upsell', enhanceUpsell],
-        ['init-follow', enhanceFollow],
     ];
     catchErrorsWithContext(modules);
 };
