+++ static/src/javascripts/projects/common/modules/identity/follow.js
@@ -1,7 +1,7 @@
 // @flow
 import React, { render } from 'preact-compat';
 import fastdom from 'lib/fastdom-promise';
-import { FollowButtonWrap } from './follow/FollowButtonWrap';
+import { FollowButtonWrap } from './upsell/follow/FollowButtonWrap';
 import loadEnhancers from './modules/loadEnhancers';
 
 const bindFollow = (el): void => {
