+++ static/src/javascripts/projects/common/modules/identity/upsell/consent-card/FollowCard.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'preact-compat';
-import { FollowButtonWrap } from 'common/modules/identity/follow/FollowButtonWrap';
+import { FollowButtonWrap } from 'common/modules/identity/upsell/follow/FollowButtonWrap';
 import type { Consent } from '../store/types';
 
 export type FollowCardProps = {
