+++ static/src/stylesheets/head.identity.scss
@@ -26,6 +26,5 @@
 @import 'module/identity/switches';
 @import 'module/identity/ad-prefs';
 @import 'module/identity/consent';
-@import 'module/identity/follow';
 @import 'module/identity/upsell/_base';
 @import 'module/identity/upsell/account-creation';
